Move to_string_view to common.h
