+++ include/spdlog/logger.h
@@ -17,7 +17,6 @@
 #include <spdlog/common.h>
 #include <spdlog/details/log_msg.h>
 #include <spdlog/details/backtracer.h>
-#include <spdlog/details/fmt_helper.h>
 
 #ifdef SPDLOG_WCHAR_TO_UTF8_SUPPORT
 #    ifndef _WIN32
@@ -88,7 +87,7 @@ public:
     template<typename... Args>
     void log(source_loc loc, level::level_enum lvl, format_string_t<Args...> fmt, Args &&... args)
     {
-        log_(loc, lvl, details::fmt_helper::to_string_view(fmt), std::forward<Args>(args)...);
+        log_(loc, lvl, details::to_string_view(fmt), std::forward<Args>(args)...);
     }
 
     template<typename... Args>
@@ -181,7 +180,7 @@ public:
     template<typename... Args>
     void log(source_loc loc, level::level_enum lvl, wformat_string_t<Args...> fmt, Args &&... args)
     {
-        log_(loc, lvl, details::fmt_helper::to_string_view(fmt), std::forward<Args>(args)...);
+        log_(loc, lvl, details::to_string_view(fmt), std::forward<Args>(args)...);
     }
 
     template<typename... Args>
@@ -374,6 +373,7 @@ protected:
 #else
             fmt::vformat_to(fmt::appender(buf), fmt, fmt::make_format_args(std::forward<Args>(args)...));
 #endif
+
             details::log_msg log_msg(loc, name_, lvl, string_view_t(buf.data(), buf.size()));
             log_it_(log_msg, log_enabled, traceback_enabled);
         }
