+++ tests/test_pattern_formatter.cpp
@@ -2,7 +2,7 @@
 #include "test_sink.h"
 
 using spdlog::memory_buf_t;
-using spdlog::details::fmt_helper::to_string_view;
+using spdlog::details::to_string_view;
 
 // log to str and return it
 template<typename... Args>
